Move isomorphic to anf_ir_utils
