+++ myia/debug/utils.py
@@ -2,9 +2,7 @@
 
 from collections import defaultdict
 from myia.debug.label import short_labeler
-from myia.anf_ir_utils import \
-    is_constant, is_parameter, is_apply, is_constant_graph, \
-    succ_deeper, succ_incoming
+from myia.anf_ir_utils import succ_deeper
 from myia.graph_utils import dfs, always_include
 
 
@@ -21,91 +19,6 @@ def mixin(target):
             setattr(target, mthd, getattr(cls, mthd))
         return target
     return apply
-
-
-def _same_node_shallow(n1, n2, equiv):
-    # Works for Constant, Parameter and nodes previously seen
-    if n1 in equiv and equiv[n1] is n2:
-        return True
-    elif is_constant_graph(n1) and is_constant_graph(n2):
-        # Note: we provide current equiv so that nested graphs can properly
-        # match their free variables, using the equiv of their parent graph.
-        return isomorphic(n1.value, n2.value, equiv)
-    elif is_constant(n1):
-        return n1.value == n2.value
-    elif is_parameter(n1):
-        return False
-    else:
-        raise TypeError(n1)
-
-
-def _same_node(n1, n2, equiv):
-    # Works for Apply (when not seen previously) or other nodes
-    if is_apply(n1):
-        return all(_same_node_shallow(i1, i2, equiv)
-                   for i1, i2 in zip(n1.inputs, n2.inputs))
-    else:
-        return _same_node_shallow(n1, n2, equiv)
-
-
-def _same_subgraph(root1, root2, equiv):
-    # Check equivalence between two subgraphs, starting from root1 and root2,
-    # using the given equivalence dictionary. This is a modified version of
-    # toposort that walks the two graphs in lockstep.
-
-    done = set()
-    todo = [(root1, root2)]
-
-    while todo:
-        n1, n2 = todo[-1]
-        if n1 in done:
-            todo.pop()
-            continue
-        cont = False
-
-        s1 = list(succ_incoming(n1))
-        s2 = list(succ_incoming(n2))
-        if len(s1) != len(s2):
-            return False
-        for i, j in zip(s1, s2):
-            if i not in done:
-                todo.append((i, j))
-                cont = True
-
-        if cont:
-            continue
-        done.add(n1)
-
-        res = _same_node(n1, n2, equiv)
-        print(res, n1, n2, equiv)
-        if res:
-            equiv[n1] = n2
-        else:
-            return False
-
-        todo.pop()
-
-    return True
-
-
-def isomorphic(g1, g2, equiv=None):
-    """Return whether g1 and g2 are structurally equivalent.
-
-    Constants are isomorphic iff they contain the same value or are isomorphic
-    graphs.
-
-    g1.return_ and g2.return_ must represent the same node under the
-    isomorphism. Parameters must match in the same order.
-    """
-    if len(g1.parameters) != len(g2.parameters):
-        return False
-
-    prev_equiv = equiv
-    equiv = dict(zip(g1.parameters, g2.parameters))
-    if prev_equiv:
-        equiv.update(prev_equiv)
-
-    return _same_subgraph(g1.return_, g2.return_, equiv)
 
 
 class GraphIndex:
